# services/api-gateway/docs/swagger.yaml
basePath: /
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  model.AdminForAdminList:
    properties:
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      phone:
        type: string
      role:
        type: string
      second_name:
        type: string
      surname:
        type: string
    type: object
...
  model.DoctorWithSpecs:
    properties:
      education:
        type: string
      email:
        type: string
      experience:
        type: integer
      firstName:
        type: string
      gender:
        type: string
      phone:
        type: string
      secondName:
        type: string
      specialty:
        items:
          type: integer
        type: array
      surname:
        type: string
      user_id:
        type: integer
    type: object
...
  model.Material:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
    type: object
...
  model.Service:
    properties:
      category_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
    type: object
  model.Spec:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: ПО для стоматологической клиники
  title: Примастом API-Gateway
  version: "1.0"
paths:
...
  /api/materials:
    post:
      consumes:
      - application/json
      description: Добавляет новый материал в систему
      parameters:
      - description: Информация о материале
        in: body
        name: material
        required: true
        schema:
          $ref: '#/definitions/model.Material'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Добавить новый материал
      tags:
      - Администратор
  /api/materials/{id}:
    delete:
      description: Удаляет материал по ID
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Удалить материал
      tags:
      - Администратор
    put:
      consumes:
      - application/json
      description: Обновляет данные материала по ID
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные материала
        in: body
        name: material
        required: true
        schema:
          $ref: '#/definitions/model.Material'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Обновить материал
      tags:
      - Администратор
...
  /api/save-admin:
    put:
      consumes:
      - application/json
      description: Обновляет информацию об администраторе
      parameters:
      - description: Данные администратора
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/model.AdminForAdminList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Некорректный ввод
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Обновить данные администратора
      tags:
      - Администратор
  /api/save-doctor:
    put:
      consumes:
      - application/json
      description: Обновляет информацию о докторе и его специализациях
      parameters:
      - description: Данные доктора
        in: body
        name: doctor
        required: true
        schema:
          $ref: '#/definitions/model.DoctorWithSpecs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Некорректный ввод
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Доступ запрещён
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Обновить данные доктора
      tags:
      - Администратор
...
  /api/services:
    post:
      consumes:
      - application/json
      description: Добавляет новую услугу в систему
      parameters:
      - description: Информация об услуге
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/model.Service'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Добавить новую услугу
      tags:
      - Администратор
  /api/services/{id}:
    put:
      consumes:
      - application/json
      description: Обновляет данные услуги по ID
      parameters:
      - description: ID услуги
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные услуги
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/model.Service'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Обновить услугу
      tags:
      - Администратор
  /api/specialties:
    get:
      description: Возвращает все доступные специализации
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Spec'
            type: array
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Получить список специализаций
      tags:
      - Администратор
  /api/staff-admin:
    get:
      description: Возвращает список всех администраторов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.AdminForAdminList'
            type: array
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Получить список администраторов
      tags:
      - Администратор
  /api/staff-doctors:
    get:
      description: Возвращает всех докторов с их специализациями
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DoctorWithSpecs'
            type: array
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Получить список докторов
      tags:
      - Администратор
...
swagger: "2.0"